+++ db/store_image_hashes.py
